add: validation with customer regist

+++ frontend/src/hooks/useReservables.ts
@@ -34,24 +34,42 @@ export const useReservables = () => {
   // フェッチしたデータで更新
   const handleRegister = async (name: string, sex: "MAN" | "WOMAN") => {
     try {
+      if (!name) {
+        window.alert("お客様のお名前は必須項目です");
+        return;
+      }
       // 実行
       await reservableAPI.regReservables(name, sex);
       // 更新後を取得
       const data = await fetchReservables();
       setReservables(data);
+      alert("登録完了");
     } catch (error) {
       console.error("登録に失敗しました", error);
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      const ApiError = error as any;
+      const message = ApiError.response?.data?.message || 
+      "登録に失敗しました";
+      alert(message);
+
     }
   };
 
   // 削除
   // フェッチしたデータで更新
   const deleteRegister = async (id: number) => {
+    if (!window.confirm("本当に削除しますか？")) {
+      return;
+    }
     try {
       await reservableAPI.deleteReservable(id);
+
       const data = await fetchReservables();
       setReservables(data);
+
+      alert("削除完了");
     } catch (error) {
+      alert("削除に失敗しました");
       console.error("削除に失敗しました", error);
     }
   };
@@ -67,8 +85,10 @@ export const useReservables = () => {
       await reservableAPI.updateReservable(id, name, sex);
       const data = await fetchReservables();
       setReservables(data);
+      alert(`更新完了`);
     } catch (error) {
       console.error("更新に失敗しました", error);
+      alert("エラー：登録情報を更新できませんでした");
     }
   };
